i fucked it all up 💀 . should be fine now tho ig

+++ CloudManager.h
@@ -9,12 +9,12 @@ namespace fdm
 	{
 	public:
 		inline static const float updateDelay = NULL; 
-		float lastUpdateTime; 
+		double lastUpdateTime; 
 		glm::vec3 offset; // 0x8
 		glm::vec3 velocity; // 0x14
 		uint32_t seed; // 0x20
 		PAD(0x4);
-		std::unordered_map<glm::i64vec3,std::unique_ptr<CloudChunk>> chunks; // 0x28
+		std::unordered_map<glm::i64vec3, std::unique_ptr<CloudChunk>> chunks; // 0x28
 		std::atomic_bool chunksReady; // 0x68
 		PAD(0x3);
 		std::atomic_int loadedChunks; // 0x6C
